+++ crates/miden-testing/src/tx_context/builder.rs
@@ -12,6 +12,7 @@ use miden_objects::EMPTY_WORD;
 use miden_objects::account::{
     Account,
     AccountHeader,
+    AccountId,
     PartialAccount,
     PartialStorage,
     PartialStorageMap,
@@ -75,11 +76,10 @@ pub type MockAuthenticator = BasicAuthenticator<ChaCha20Rng>;
 pub struct TransactionContextBuilder {
     source_manager: Arc<dyn SourceManagerSync>,
     account: Account,
-    account_seed: Option<Word>,
     advice_inputs: AdviceInputs,
     authenticator: Option<MockAuthenticator>,
     expected_output_notes: Vec<Note>,
-    foreign_account_inputs: Vec<AccountInputs>,
+    foreign_account_inputs: BTreeMap<AccountId, AccountInputs>,
     input_notes: Vec<Note>,
     tx_script: Option<TransactionScript>,
     tx_script_args: Word,
@@ -87,7 +87,7 @@ pub struct TransactionContextBuilder {
     transaction_inputs: Option<TransactionInputs>,
     auth_args: Word,
     signatures: Vec<(PublicKey, Word, Vec<Felt>)>,
-    load_partial_account: bool,
+    is_lazy_loading_enabled: bool,
 }
 
 impl TransactionContextBuilder {
@@ -95,7 +95,6 @@ impl TransactionContextBuilder {
         Self {
             source_manager: Arc::new(DefaultSourceManager::default()),
             account,
-            account_seed: None,
             input_notes: Vec::new(),
             expected_output_notes: Vec::new(),
             tx_script: None,
@@ -104,10 +103,10 @@ impl TransactionContextBuilder {
             advice_inputs: Default::default(),
             transaction_inputs: None,
             note_args: BTreeMap::new(),
-            foreign_account_inputs: vec![],
+            foreign_account_inputs: BTreeMap::new(),
             auth_args: EMPTY_WORD,
             signatures: Vec::new(),
-            load_partial_account: false,
+            is_lazy_loading_enabled: false,
         }
     }
 
@@ -147,12 +146,6 @@ impl TransactionContextBuilder {
         Self::new(account)
     }
 
-    /// Override and set the account seed manually
-    pub fn account_seed(mut self, account_seed: Option<Word>) -> Self {
-        self.account_seed = account_seed;
-        self
-    }
-
     /// Extend the advice inputs with the provided [AdviceInputs] instance.
     pub fn extend_advice_inputs(mut self, advice_inputs: AdviceInputs) -> Self {
         self.advice_inputs.extend(advice_inputs);
@@ -175,8 +168,9 @@ impl TransactionContextBuilder {
     }
 
     /// Set foreign account codes that are used by the transaction
-    pub fn foreign_accounts(mut self, inputs: Vec<AccountInputs>) -> Self {
-        self.foreign_account_inputs = inputs;
+    pub fn foreign_accounts(mut self, inputs: impl IntoIterator<Item = AccountInputs>) -> Self {
+        self.foreign_account_inputs
+            .extend(inputs.into_iter().map(|account_inputs| (account_inputs.id(), account_inputs)));
         self
     }
 
@@ -216,11 +210,13 @@ impl TransactionContextBuilder {
     }
 
     /// Causes the transaction to only construct a minimal partial account as the transaction
-    /// input, causing lazy loading of assets throughout transaction execution.
+    /// input, causing lazy loading of assets and storage map items throughout transaction
+    /// execution. Additionally, foreign accounts aren't provided via the transaction args but are
+    /// lazy loaded as well.
     ///
     /// This exists to test lazy loading selectively and should go away in the future.
-    pub fn enable_partial_loading(mut self) -> Self {
-        self.load_partial_account = true;
+    pub fn enable_lazy_loading(mut self) -> Self {
+        self.is_lazy_loading_enabled = true;
         self
     }
 
@@ -283,10 +279,9 @@ impl TransactionContextBuilder {
 
                 let input_note_ids: Vec<NoteId> =
                     mock_chain.committed_notes().values().map(MockChainNote::id).collect();
-                let account = PartialAccount::from(&self.account);
 
                 mock_chain
-                    .get_transaction_inputs(account, self.account_seed, &input_note_ids, &[])
+                    .get_transaction_inputs(&self.account, &input_note_ids, &[])
                     .context("failed to get transaction inputs from mock chain")?
             },
         };
@@ -294,9 +289,8 @@ impl TransactionContextBuilder {
         // If partial loading is enabled, construct an account that doesn't contain all
         // merkle paths of assets and storage maps, in order to test lazy loading.
         // Otherwise, load the full account.
-        let tx_inputs = if self.load_partial_account {
-            let (account, account_seed, block_header, partial_blockchain, input_notes) =
-                tx_inputs.into_parts();
+        let tx_inputs = if self.is_lazy_loading_enabled {
+            let (account, block_header, partial_blockchain, input_notes) = tx_inputs.into_parts();
             // Construct a partial vault that tracks the empty word, but none of the assets
             // that are actually in the asset tree. That way, the partial vault has the same
             // root as the full vault, but will not add any relevant merkle paths to the
@@ -333,20 +327,21 @@ impl TransactionContextBuilder {
                 account.code().clone(),
                 partial_storage,
                 partial_vault,
-            );
+                None,
+            )?;
 
-            TransactionInputs::new(
-                account,
-                account_seed,
-                block_header,
-                partial_blockchain,
-                input_notes,
-            )?
+            TransactionInputs::new(account, block_header, partial_blockchain, input_notes)?
         } else {
             tx_inputs
         };
 
-        let tx_args = TransactionArgs::new(AdviceMap::default(), self.foreign_account_inputs)
+        let foreign_account_inputs = if self.is_lazy_loading_enabled {
+            Vec::new()
+        } else {
+            self.foreign_account_inputs.values().cloned().collect()
+        };
+
+        let tx_args = TransactionArgs::new(AdviceMap::default(), foreign_account_inputs)
             .with_note_args(self.note_args);
 
         let mut tx_args = if let Some(tx_script) = self.tx_script {
@@ -368,7 +363,7 @@ impl TransactionContextBuilder {
             let mast_forest_store = TransactionMastStore::new();
             mast_forest_store.load_account_code(tx_inputs.account().code());
 
-            for acc_inputs in tx_args.foreign_account_inputs() {
+            for acc_inputs in self.foreign_account_inputs.values() {
                 mast_forest_store.insert(acc_inputs.code().mast());
             }
 
@@ -378,6 +373,7 @@ impl TransactionContextBuilder {
         Ok(TransactionContext {
             account: self.account,
             expected_output_notes: self.expected_output_notes,
+            foreign_account_inputs: self.foreign_account_inputs,
             tx_args,
             tx_inputs,
             mast_store,